Anda mef, falta hora

+++ STM32F103C6/delay.c
@@ -1,8 +1,8 @@
 #include "delay.h"
-void delay_us (uint16_t t)
+void delay_us (uint32_t t)
 {
    volatile unsigned long l = 0;
-   uint16_t i;
+   uint32_t i;
    for( i= 0; i < t; i++)
    for(l = 0; l < 6; l++)
    { }
