+++ STM32F103C6/delay.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
-void delay_us (uint16_t );
+void delay_us (uint32_t );
